bondcorrel: Save pairs in sets
Previously, bond connectivity information was saved to a list of lists
with 0 indicating no bond, and 1 indicating a bond (a "sparse matrix"
so to speak). However, naturally this required a prohibitively large
amount of memory. Therefore, now a set with pairs of size_t saves
just the connected pairs ("dense matrix"). The set enables fast lookup
as well. There is now some loss of generality, since this can only be
used for connectivity information. Also some templated functions could
be dispensed with. Tests were modified to reflect this change.