bondcorrel: Template on the Network type
The `time_correlation_function` now takes a vector of UndirectedNetwork
or DirectedNetwork objects. Therefore this function now has to be
templated on the Network type (just templating on double does not
work because it cannot infer the type). The other option would have been
to take a vector of pointers, but this required the smallest changes to
work. The tests were modified to accomodate this.